Fix build with Visual Studio 17 2022

+++ modules/core/src/math/misc/vpMath.cpp
@@ -509,7 +509,7 @@ std::vector<std::pair<double, double> > vpMath::computeRegularPointsOnSphere(uns
   double d_phi = a / d_theta;
 
   std::vector<std::pair<double, double> > lonlat_vec;
-  lonlat_vec.reserve(std::sqrt(maxPoints));
+  lonlat_vec.reserve(static_cast<unsigned int>(std::sqrt(maxPoints)));
   for (int m = 0; m < m_theta; m++) {
     double theta = M_PI * (m + 0.5) / m_theta;
     int m_phi = static_cast<int>(round(2.0 * M_PI * sin(theta) / d_phi));
